Updared to new VWSI api.

+++ Sources/Renderer/RenderStage.hpp
@@ -34,7 +34,7 @@ namespace fl
 
 		void Rebuild(Swapchain *swapchain);
 
-		uint32_t SubpassCount() const { return m_renderpassCreate->GetSubpasses().size(); };
+		uint32_t SubpassCount() const { return static_cast<uint32_t>(m_renderpassCreate->GetSubpasses().size()); };
 
 		uint32_t GetWidth() const;
